Rename actions to commands
