+++ lua/wtf/commands/history.lua
