+++ lua/wtf/commands/search.lua
